Switch to reference resolution
Instead of binding the game to the size of the window it makes more
sense to use a fixed resolution and define te world in terms of it. The
orthographic projection matrix takes care of scaling the game to the
correct size to fill the window.

This allows the game to use a pixels-per-metre conversion to allow for a
more natural description of the world.